detect previously exported libraries

+++ aperture.py
@@ -572,21 +572,33 @@ class Aplib():
                 + str(datetime.now().__format__("%Y%m%d%H%M%S")))
         self.name_of["TopLevelAlbums"] = "Albums"
 
-        #create temporary file
-        tmp_file = Path(self.export_path) / (self.name_of[root_uuid] + ".inprogress")
-        with open(tmp_file, "w") as f:
-            f.write("")
 
-        try:
-            #Do the export
-            self.recursiveExport(root_uuid, self.export_path)
-            #delete temporary file when finished
-            os.remove(tmp_file)
+        already_exported = False
+        for _,dirs,_ in os.walk(self.export_path):
+            for d in dirs:
+                if (self.path_to_aplib.name + " xptd ") in d:
+                    already_exported = True
+                    break
 
-        except Exception as e:
-            print("Exception!")
-            print(self.path_to_aplib)
-            print(e)
+
+        if not already_exported:
+            #create temporary file
+            tmp_file = Path(self.export_path) / (self.name_of[root_uuid] + ".inprogress")
+            with open(tmp_file, "w") as f:
+                f.write("")
+
+            try:
+                #Do the export
+                self.recursiveExport(root_uuid, self.export_path)
+                #delete temporary file when finished
+                os.remove(tmp_file)
+
+            except Exception as e:
+                print("Exception!")
+                print(self.path_to_aplib)
+                print(e)
+        else:
+            print("Already exported")
